requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
channels==4.0.0
channels-redis==4.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
msgpack==1.0.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
